Drive LEDs from timer interrupt
This decouples the LED logic from the USB polling requirements.
The timer is set to fire at 200 Hz, every 5 ms. The USB requirement
is a poll rate of at least 100 Hz, every 10 ms.

To ensure the LED isn't driven too fast, the ISR is subsampled
by a factor of 6 to achieve a rate of 33 Hz, every 30 ms.

Signed-off-by: Markus Mayer <widemeadows@gmail.com>